shared ui 구조 변경
