+++ src/shared/ui/card/ui/card-header.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+
+export const CardHeader = forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
+  ({ className, ...props }, ref) => (
+    <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+  ),
+)
+CardHeader.displayName = "CardHeader"
