+++ src/shared/ui/dialog/ui/index.ts
@@ -1,0 +1,10 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export { DialogContent } from "./dialog-content"
+export { DialogHeader } from "./dialog-header"
+export { DialogTitle } from "./dialog-title"
+
+export const Dialog = DialogPrimitive.Root
+export const DialogTrigger = DialogPrimitive.Trigger
+export const DialogPortal = DialogPrimitive.Portal
+export const DialogOverlay = DialogPrimitive.Overlay
