+++ src/shared/ui/select/ui/index.ts
@@ -1,0 +1,9 @@
+import * as SelectPrimitive from "@radix-ui/react-select"
+
+export { SelectContent } from "./select-content"
+export { SelectItem } from "./select-item"
+export { SelectTrigger } from "./select-trigger"
+
+export const Select = SelectPrimitive.Root
+export const SelectGroup = SelectPrimitive.Group
+export const SelectValue = SelectPrimitive.Value
