+++ src/shared/ui/table/ui/index.ts
@@ -1,0 +1,6 @@
+export { Table } from "./table"
+export { TableBody } from "./table-body"
+export { TableCell } from "./table-cell"
+export { TableHead } from "./table-head"
+export { TableHeader } from "./table-header"
+export { TableRow } from "./table-row"
