+++ src/shared/ui/table/ui/table-body.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+interface TableBodyProps extends React.TableHTMLAttributes<HTMLTableSectionElement> {
+  className?: string
+}
+
+export const TableBody = forwardRef<HTMLTableSectionElement, TableBodyProps>(({ className, ...props }, ref) => (
+  <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+))
+TableBody.displayName = "TableBody"
